set scheduler unit tests and rename
